pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "Django E-Commerce project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.4",
    "django-cors-headers>=4.6.0",
    "django-dotenv>=1.4.2",
    "django-environ>=0.12.0",
    "django-oauth-toolkit>=3.0.1",
    "django-silk>=5.3.2",
    "djangorestframework>=3.15.2",
    "djangorestframework-camel-case>=1.4.2",
    "djoser>=2.3.1",
    "drf-spectacular[sidecar]>=0.28.0",
    "whitenoise>=6.9.0",
    "uvicorn>=0.34.0",

]

[dependency-groups]
dev = [
    "autopep8>=2.3.1",
    "granian>=1.7.5",
    "ngrok>=1.4.0",
    "pytest>=8.3.4",
    "django-debug-toolbar>=5.0.1",

]
